Added stylesheetPath param
Added a parameter to allow the caller to specify a stylesheetPath for
the JUnitXmlReporter that will add an xml-stylesheet tag to the xml file
so that it can be opened directly in a browser.  Note: Chrome does not
support this if both files are local.  It considers it a CORS violation
for some reason.