Adds system-out callback option
This change allows the addition of a `<system-out></system-out>` block to each `<testcase>`. This is useful for adding extra information that can be used in CI systems like Jenkins / Hudson.

For example, using something like protractor-screenshot-reporter, this allows us to add `[[ATTACHMENT|spec/as/filename/description.png]]` for use in the JUnit Attachments Plugin.